REFACTOR: promo.tsx 컴포넌트 기반 구조로 분리
promo.tsx 파일을 1,279줄에서 65줄로 축소하고 컴포넌트 기반 구조로 리팩토링:
- 상수 및 타입 정의를 promo/constants.ts로 분리
- SVG 아이콘을 promo/icons.tsx로 분리
- 9개의 섹션 컴포넌트를 개별 파일로 분리
  - HeaderSection, HeroSection, PartnersSection
  - PortfolioSection, FeaturesSection, MusesSection
  - FaqSection, ContactBar, FooterSection

기존: 단일 파일 1,279줄
변경 후: 메인 파일 65줄 + 11개 모듈 파일
