+++ apps/shop/src/routes/promo/icons.tsx
@@ -1,0 +1,59 @@
+/**
+ * InstagramIcon - Instagram platform SVG icon
+ */
+export function InstagramIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+    >
+      <path
+        d="M10.0001 5.71755C12.3612 5.71755 14.3056 7.63885 14.3056 9.99996C14.3056 12.3611 12.3612 14.3055 10.0001 14.3055C7.63897 14.3055 5.71767 12.3611 5.71767 9.99996C5.71767 7.63885 7.63897 5.71755 10.0001 5.71755ZM10.0001 12.7777C11.5279 12.7777 12.7779 11.5277 12.7779 9.99996C12.7779 8.47218 11.5279 7.22218 10.0001 7.22218C8.4723 7.22218 7.2223 8.47218 7.2223 9.99996C7.2223 11.5277 8.4723 12.7777 10.0001 12.7777ZM14.4445 6.55089C13.9121 6.55089 13.4723 6.08792 13.4723 5.55552C13.4723 4.99996 13.9121 4.56015 14.4445 4.56015C15.0001 4.56015 15.463 4.99996 15.463 5.55552C15.463 6.08792 15.0001 6.55089 14.4445 6.55089ZM10.0001 1.66663C12.2686 1.66663 12.5464 1.68978 13.4492 1.71292C14.3288 1.75922 14.9538 1.89811 15.4862 2.10644C16.0186 2.31477 16.4816 2.6157 16.9445 3.05551C17.4075 3.54163 17.6853 3.98144 17.8936 4.537C18.1019 5.0694 18.264 5.6944 18.2871 6.57403C18.3103 7.45366 18.3334 7.73144 18.3334 9.99996C18.3334 12.2685 18.3103 12.5463 18.2871 13.449C18.264 14.3055 18.1019 14.9537 17.8936 15.4629C17.6853 16.0185 17.3843 16.4814 16.9445 16.9444C16.4816 17.3842 16.0186 17.6851 15.463 17.8935C14.9538 18.1018 14.3056 18.2638 13.4492 18.287C12.5464 18.3101 12.2686 18.3333 10.0001 18.3333C7.73156 18.3333 7.45379 18.3101 6.57416 18.287C5.69453 18.2638 5.06953 18.1018 4.53712 17.8935C3.98156 17.6851 3.54175 17.3842 3.05564 16.9444C2.61582 16.4814 2.3149 16.0185 2.10656 15.4629C1.89823 14.9537 1.73619 14.3055 1.71304 13.449C1.6899 12.5463 1.66675 12.2685 1.66675 9.99996C1.66675 7.73144 1.6899 7.45366 1.71304 6.55089C1.73619 5.6944 1.89823 5.0694 2.10656 4.537C2.3149 3.98144 2.61582 3.54163 3.05564 3.05551C3.54175 2.6157 3.98156 2.31477 4.53712 2.10644C5.06953 1.89811 5.69453 1.73607 6.57416 1.71292C7.45379 1.68978 7.73156 1.66663 10.0001 1.66663ZM10.0001 3.17126L6.62045 3.1944C5.83342 3.26385 5.3936 3.37959 5.09267 3.49533C4.69916 3.65737 4.42138 3.84255 4.12045 4.12033C3.84267 4.39811 3.65749 4.67589 3.49545 5.0694C3.37971 5.37033 3.26397 5.81014 3.21767 6.62033L3.19453 9.99996L3.21767 13.3796C3.26397 14.1666 3.37971 14.6064 3.49545 14.9074C3.65749 15.3009 3.84267 15.5787 4.12045 15.8796C4.42138 16.1574 4.69916 16.3426 5.09267 16.4814C5.3936 16.5972 5.83342 16.7361 6.62045 16.7824L10.0001 16.8055L13.3797 16.7824C14.1899 16.7361 14.6297 16.5972 14.9306 16.4814C15.3242 16.3426 15.6019 16.1574 15.8797 15.8796C16.1575 15.5787 16.3427 15.3009 16.5047 14.9074C16.6205 14.6064 16.7593 14.1666 16.8056 13.3796L16.8288 9.99996L16.8056 6.62033C16.7593 5.83329 16.6205 5.37033 16.5047 5.0694C16.3427 4.67589 16.1575 4.39811 15.8797 4.12033C15.6019 3.84255 15.3242 3.65737 14.9306 3.49533C14.6297 3.37959 14.1899 3.26385 13.3797 3.1944L10.0001 3.17126Z"
+        fill="#18181B"
+      />
+    </svg>
+  );
+}
+
+/**
+ * TiktokIcon - TikTok platform SVG icon
+ */
+export function TiktokIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+    >
+      <path
+        d="M13.3542 1.66663C13.3542 3.79163 15 5.54163 17.125 5.60413V8.45829C15.75 8.43746 14.4583 7.95829 13.3958 7.20829V13.0416C13.3958 15.9375 11.0417 18.3333 8.14583 18.3333C5.22917 18.3333 2.875 15.9583 2.875 13.0416C2.875 10.1041 5.1875 7.74996 8.04167 7.70829V10.5416C6.75 10.6041 5.6875 11.7083 5.6875 13.0416C5.6875 14.375 6.77083 15.4791 8.125 15.4791C9.45833 15.4791 10.5417 14.375 10.5417 13.0416V1.66663H13.3542Z"
+        fill="#18181B"
+      />
+    </svg>
+  );
+}
+
+/**
+ * YoutubeIcon - YouTube platform SVG icon
+ */
+export function YoutubeIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+    >
+      <path
+        d="M17.7 5.9C17.5 5.2 16.9 4.6 16.2 4.5C14.8 4.2 10 4.2 10 4.2C10 4.2 5.2 4.2 3.8 4.5C3.1 4.7 2.5 5.2 2.3 5.9C2 7.3 2 10 2 10C2 10 2 12.7 2.3 14.1C2.5 14.8 3.1 15.4 3.8 15.5C5.2 15.8 10 15.8 10 15.8C10 15.8 14.8 15.8 16.2 15.5C16.9 15.3 17.5 14.8 17.7 14.1C18 12.7 18 10 18 10C18 10 18 7.3 17.7 5.9ZM8.3 12.5V7.5L12.7 10L8.3 12.5Z"
+        fill="#18181B"
+      />
+    </svg>
+  );
+}
